# swagger.yaml
openapi: 3.0.3
info:
  title: Bell Project API
  description: QR-based restaurant ordering system API
  version: 1.0.0
  contact:
    name: Bell Project Team
    email: support@bellproject.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.bellproject.com/dev
    description: Development server
  - url: https://api.bellproject.com/prod
    description: Production server
...
paths:
  /menu:
    get:
      tags:
        - menu
      summary: Get restaurant menu
      description: Retrieve the latest confirmed menu for a restaurant
      parameters:
        - name: restaurantId
          in: query
          required: true
          description: Unique restaurant identifier
          schema:
            type: string
            example: "rest_001"
      responses:
        '200':
          description: Menu retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMenuResponse'
        '400':
          description: Bad request - Restaurant ID is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Menu not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - menu
      summary: Create or update menu
      description: Create a new menu version or update existing menu
      parameters:
        - name: restaurantId
          in: query
          required: true
          description: Unique restaurant identifier
          schema:
            type: string
            example: "rest_001"
        - name: confirm
          in: query
          required: false
          description: Whether to confirm the menu immediately
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostMenuRequest'
      responses:
        '201':
          description: Menu created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostMenuResponse'
        '400':
          description: Bad request - Invalid menu data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Menu version already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    MenuItem:
      type: object
      required:
        - id
        - name
        - price
        - category
        - available
      properties:
        id:
          type: string
          example: "item_001"
        name:
          type: string
          example: "Classic Burger"
        price:
          type: number
          format: float
          minimum: 0
          example: 12.99
        description:
          type: string
          example: "Juicy beef patty with lettuce, tomato, and onion"
        category:
          type: string
          example: "burgers"
        available:
          type: boolean
          example: true
        image:
          type: string
          format: uri
          example: "https://example.com/burger.jpg"
        options:
          type: array
          items:
            $ref: '#/components/schemas/MenuItemOption'
    MenuItemOption:
      type: object
      required:
        - id
        - name
        - required
        - choices
      properties:
        id:
          type: string
          example: "size"
        name:
          type: string
          example: "Size"
        required:
          type: boolean
          example: true
        choices:
          type: array
          items:
            $ref: '#/components/schemas/OptionChoice'
    OptionChoice:
      type: object
      required:
        - id
        - name
        - priceModifier
      properties:
        id:
          type: string
          example: "large"
        name:
          type: string
          example: "Large"
        priceModifier:
          type: number
          format: float
          example: 2.00
...
    PostMenuRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MenuItem'
          minItems: 1
    PostMenuResponse:
      type: object
      properties:
        version:
          type: string
          example: "v1692123456789"
        status:
          type: string
          enum: [DRAFT, CONFIRMED]
          example: "DRAFT"
        createdAt:
          type: string
          format: date-time
          example: "2023-08-15T12:30:45.123Z"
    GetMenuResponse:
      type: object
      properties:
        menu:
          type: object
          properties:
            restaurantId:
              type: string
              example: "rest_001"
            version:
              type: string
              example: "v1692123456789"
            items:
              type: array
              items:
                $ref: '#/components/schemas/MenuItem'
            status:
              type: string
              enum: [DRAFT, CONFIRMED]
              example: "CONFIRMED"
            createdAt:
              type: string
              format: date-time
              example: "2023-08-15T12:30:45.123Z"
            confirmedAt:
              type: string
              format: date-time
              example: "2023-08-15T12:35:45.123Z"
...
    ErrorResponse:
      type: object
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
          example: "https://httpstatuses.com/400"
        title:
          type: string
          example: "Bad Request"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "The request is missing required parameters"
        instance:
          type: string
          format: uri
          example: "/menu"